requirements.txt
websocket-client==1.6.3
websockets==12.0
numpy>=1.26.0
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.0
Flask==2.3.3
Flask-SocketIO==5.3.4
python-socketio==5.8.0
python-engineio==4.7.0
eventlet==0.33.3
python-dotenv==1.0.0
requests==2.31.0